requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==1.11.18
django-geojson==2.12.0
django-hashedfilenamestorage==2.4.1
django-imagefield==0.7.3
django-leaflet==0.24.0
django-multiselectfield==0.1.8
django-reversion==3.0.3
djangorestframework==3.9.1
djangorestframework-gis==0.14
ExifRead==2.1.2
gunicorn==19.9.0
idna==2.8
mysql-connector==2.1.6
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pygeoif==0.7
pyshp==2.0.1
pytz==2018.9
requests==2.21.0
six==1.12.0
sorl-thumbnail==12.5.0
sorl-thumbnail-serializer-field==0.2.1
urllib3==1.24.2
